:recycle: refactor: kafka 토픽 이벤트 수정
