+++ src/main/java/com/mulmeong/contest/read/application/ContestPostServiceImpl.java
@@ -4,8 +4,8 @@ import com.mulmeong.contest.read.common.exception.BaseException;
 import com.mulmeong.contest.read.common.response.BaseResponseStatus;
 import com.mulmeong.contest.read.entity.ContestPost;
 import com.mulmeong.contest.read.infrastructure.ContestPostRepository;
-import com.mulmeong.event.contest.ContestPostCreateEvent;
-import com.mulmeong.event.contest.ContestPostUpdateEvent;
+import com.mulmeong.event.contest.consume.ContestPostCreateEvent;
+import com.mulmeong.event.contest.consume.ContestVoteUpdateEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -24,7 +24,7 @@ public class ContestPostServiceImpl implements ContestPostService{
     }
 
     @Override
-    public void updateContestPost(ContestPostUpdateEvent message) {
+    public void updateContestVote(ContestVoteUpdateEvent message) {
 
         ContestPost post = contestPostRepository.findByPostUuid(message.getPostUuid()).orElseThrow(
                 () -> new BaseException(BaseResponseStatus.NOT_EXIST_POST)
