+++ src/main/java/com/mulmeong/event/contest/consume/ContestPostCreateEvent.java
@@ -1,9 +1,8 @@
-package com.mulmeong.event.contest;
+package com.mulmeong.event.contest.consume;
 
 
 import com.mulmeong.contest.read.entity.ContestPost;
 import com.mulmeong.contest.read.entity.MediaType;
-import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
