+++ src/main/java/com/mulmeong/event/contest/consume/ContestVoteUpdateEvent.java
@@ -1,7 +1,6 @@
-package com.mulmeong.event.contest;
+package com.mulmeong.event.contest.consume;
 
 import com.mulmeong.contest.read.entity.ContestPost;
-import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,8 +8,9 @@ import lombok.NoArgsConstructor;
 @Data
 @Getter
 @NoArgsConstructor
-public class ContestPostUpdateEvent {
+public class ContestVoteUpdateEvent {
 
+    private Long contestId;
     private String postUuid;
     private Integer count;
